radar-pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "radar-pipeline"
version = "0.1.0"
description = "Radar point cloud processing and clustering pipeline"
authors = [
    {name = "Samuel Cancilla", email = "scancilla06@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "matplotlib>=3.5.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
radar-pipeline = "radar_pipeline.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/radar_pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E203"]
